@@ -16,19 +16,31 @@
         "source": "https://github.com/YOURLS/YOURLS"
     },
     "require": {
-        "php": ">=5.6",
+        "php": ">=7.4",
+        "ext-dom": "*",
+        "ext-filter": "*",
+        "ext-hash": "*",
+        "ext-pcre": "*",
         "ext-pdo": "*",
-        "ozh/bookmarkletgen": "1.2",
-        "ozh/phpass": "1.2.0",
-        "rmccue/requests" : "1.7",
-        "pomo/pomo" : "1.4.1",
-        "geoip2/geoip2" : "2.9.0",
-        "aura/sql": "~2.",
-        "jakeasmith/http_build_url": "1.0.1",
-        "symfony/polyfill-mbstring": "1.10.0"
+        "ext-pdo_mysql": "*",
+        "ozh/bookmarkletgen": "^1.2",
+        "rmccue/requests" : "^2.0",
+        "pomo/pomo" : "^1.4",
+        "geoip2/geoip2" : "^2.10",
+        "aura/sql": "^3.0",
+        "jakeasmith/http_build_url": "^1.0",
+        "symfony/polyfill-mbstring": "^1.15",
+        "symfony/polyfill-intl-idn": "^1.17",
+        "spatie/array-to-xml": "^2.14"
+    },
+    "require-dev": {
+        "ext-ctype": "*"
     },
     "config": {
-        "vendor-dir": "includes/vendor"
+        "vendor-dir": "includes/vendor",
+        "platform": {
+            "php": "7.4.0"
+        }
     },
     "autoload": {
         "psr-4": {
@@ -36,8 +48,15 @@
         }
     },
     "suggest": {
+        "ext-bcmath": "May be needed to read GeoIP database (or ext-gmp)",
+        "ext-curl": "Required for API usage",
+        "ext-gmp": "May be needed to read GeoIP database (or ext-bcmath)",
+        "ext-iconv": "For safer input handling",
+        "ext-json": "For faster API performance",
         "ext-mbstring": "For best performance",
-        "ext-curl": "Required for API usage"
+        "ext-openssl": "To fetch titles from HTTPS sites",
+        "ext-posix": "May be needed on certain PHP versions",
+        "ext-zlib": "For best performance"
     },
     "scripts": {
         "post-update-cmd": "bash ./includes/vendor/build-script/yourls-build.sh ./includes/vendor"
