@@ -24,12 +24,13 @@
     "require": {
         "ext-openssl": "*",
         "ext-pcre": "*",
-        "php": "^5.3.2 || ^7.0"
+        "php": "^5.3.2 || ^7.0 || ^8.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.8.35 || ^5.7 || ^6.5",
+        "symfony/phpunit-bridge": "^4.2 || ^5",
+        "phpstan/phpstan": "^0.12.55",
         "psr/log": "^1.0",
-        "symfony/process": "^2.5 || ^3.0 || ^4.0"
+        "symfony/process": "^2.5 || ^3.0 || ^4.0 || ^5.0 || ^6.0"
     },
     "autoload": {
         "psr-4": {
@@ -43,12 +44,11 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "1.x-dev"
+            "dev-main": "1.x-dev"
         }
     },
-    "config": {
-        "platform": {
-            "php": "5.3.9"
-        }
+    "scripts": {
+        "test": "SYMFONY_PHPUNIT_REMOVE_RETURN_TYPEHINT=1 vendor/bin/simple-phpunit",
+        "phpstan": "vendor/bin/phpstan analyse"
     }
 }
