@@ -6,10 +6,7 @@
  * @group API
  * @since 0.1
  */
-class API_Func_Tests extends PHPUnit_Framework_TestCase {
-
-    protected function tearDown() {
-    }
+class API_Func_Tests extends PHPUnit\Framework\TestCase {
 
     public function api_actions() {
         return array(
@@ -30,8 +27,11 @@ class API_Func_Tests extends PHPUnit_Framework_TestCase {
      */
     public function test_api_actions( $action, $alias ) {
         $action = $alias ? $alias : $action;
-        
-        $this->assertTrue( is_array( call_user_func( 'yourls_api_action_' . $action ) ) );
+
+        $function = 'yourls_api_action_' . $action;
+
+        $this->assertTrue( is_callable( $function ) );
+        $this->assertTrue( is_array( $function() ) );
     }
 
 }
