@@ -4,16 +4,18 @@
  * This abstract class isn't supposed to be run as tests
  * See login_*.php files
  */
-abstract class Login_Base extends PHPUnit_Framework_TestCase {
+abstract class Login_Base extends PHPUnit\Framework\TestCase {
 
     protected $backup_request;
-    
-    protected function setUp() {
+
+    protected function setUp(): void {
         $this->backup_request = $_REQUEST;
+        $_REQUEST['nonce'] = yourls_create_nonce('admin_login');
     }
 
-    protected function tearDown() {
+    protected function tearDown(): void {
         $_REQUEST = $this->backup_request;
+        yourls_remove_all_actions('pre_yourls_die');
     }
 
 	/**
@@ -25,14 +27,14 @@ abstract class Login_Base extends PHPUnit_Framework_TestCase {
         $pre_login    = yourls_did_action( 'pre_login' );
         $login        = yourls_did_action( 'login' );
         $login_failed = yourls_did_action( 'login_failed' );
-        
+
 		$this->assertTrue( yourls_is_valid_user() );
-        
+
 		$this->assertEquals( $pre_login + 1, yourls_did_action( 'pre_login' ) );
 		$this->assertEquals( $login + 1, yourls_did_action( 'login' ) );
 		$this->assertEquals( $login_failed, yourls_did_action( 'login_failed' ) );
 	}
-	
+
     /**
      * Check that auth is shuntable
      *
@@ -43,7 +45,7 @@ abstract class Login_Base extends PHPUnit_Framework_TestCase {
         $this->assertSame( array(), yourls_is_valid_user() );
         yourls_remove_filter( 'shunt_is_valid_user', 'yourls_return_empty_array' );
     }
-    
+
     /**
      * Check that auth returns false with no credential
      *
@@ -55,11 +57,11 @@ abstract class Login_Base extends PHPUnit_Framework_TestCase {
         $login_failed = yourls_did_action( 'login_failed' );
 
         $this->assertNotTrue( yourls_is_valid_user() );
-        
+
 		$this->assertEquals( $login, yourls_did_action( 'login' ) );
 		$this->assertEquals( $login_failed + 1, yourls_did_action( 'login_failed' ) );
     }
-    
+
     /**
      * Check that auth returns false with empty credential
      *
@@ -75,7 +77,7 @@ abstract class Login_Base extends PHPUnit_Framework_TestCase {
 		$this->assertEquals( $login, yourls_did_action( 'login' ) );
 		$this->assertEquals( $login_failed + 1, yourls_did_action( 'login_failed' ) );
     }
-    
+
     /**
      * Check that auth returns false with incorrect credentials
      *
@@ -86,10 +88,15 @@ abstract class Login_Base extends PHPUnit_Framework_TestCase {
         $login        = yourls_did_action( 'login' );
         $login_failed = yourls_did_action( 'login_failed' );
 
-        $this->assertNotTrue( yourls_is_valid_user() );
+        // with "normal" logins, we simulate the login forms and the presence of a nonce
+        if (get_class($this) == 'Auth_Login_Normal_Tests') {
+            $this->expectException(Exception::class);
+            $this->expectExceptionMessage('I have died');
+            // intercept yourls_die() before it actually dies
+            yourls_add_action( 'pre_yourls_die', function() { throw new Exception( 'I have died' ); } );
+        }
 
-		$this->assertEquals( $login, yourls_did_action( 'login' ) );
-		$this->assertEquals( $login_failed + 1, yourls_did_action( 'login_failed' ) );
+        $this->assertNotTrue( yourls_is_valid_user() );
     }
 
 }
