@@ -6,18 +6,18 @@
  * @group signatures
  * @since 0.1
  */
-class Auth_Sig_Tests extends PHPUnit_Framework_TestCase {
+class Auth_Sig_Tests extends PHPUnit\Framework\TestCase {
 
     protected $backup_request;
-    
-    protected function setUp() {
+
+    protected function setUp(): void {
         $this->backup_request = $_REQUEST;
     }
 
-    protected function tearDown() {
+    protected function tearDown(): void {
         $_REQUEST = $this->backup_request;
     }
-    
+
     /**
      * Check that empty signature isn't valid
      *
@@ -59,7 +59,58 @@ class Auth_Sig_Tests extends PHPUnit_Framework_TestCase {
         $_REQUEST['timestamp'] = rand_str();
         $this->assertFalse( yourls_check_signature_timestamp() );
     }
-    
+
+    /**
+     * Check that valid md5 timestamped sig is valid
+     *
+     * @since 0.1
+     */
+    public function test_signature_timestamp_md5() {
+        $timestamp = time();
+        $_REQUEST['timestamp'] = $timestamp;
+
+        global $yourls_user_passwords;
+        $random_user = array_rand($yourls_user_passwords);
+        $signature = yourls_auth_signature($random_user);
+
+        $md5 = md5( $timestamp . $signature );
+        $_REQUEST['signature'] = $md5;
+        $this->assertTrue( yourls_check_signature_timestamp() );
+
+        $md5 = md5( $signature . $timestamp );
+        $_REQUEST['signature'] = $md5;
+        $this->assertTrue( yourls_check_signature_timestamp() );
+    }
+
+    /**
+     * Check that valid hashed timestamped sig is valid
+     *
+     * @since 0.1
+     */
+    public function test_signature_timestamp_hash() {
+        $timestamp = time();
+        $_REQUEST['timestamp'] = $timestamp;
+
+        global $yourls_user_passwords;
+        $random_user = array_rand($yourls_user_passwords);
+        $signature = yourls_auth_signature($random_user);
+
+        $algos = hash_algos();
+        $random_algo = $algos[array_rand($algos)];
+        $_REQUEST['hash'] = $random_algo;
+
+        $hash = hash($random_algo, $timestamp . $signature );
+        $_REQUEST['signature'] = $hash;
+        $this->assertTrue( yourls_check_signature_timestamp() );
+
+        $hash = hash($random_algo, $signature . $timestamp );
+        $_REQUEST['signature'] = $hash;
+        $this->assertTrue( yourls_check_signature_timestamp() );
+
+        $_REQUEST['hash'] = rand_str();
+        $this->assertFalse( yourls_check_signature_timestamp() );
+    }
+
     /**
      * Provide valid and invalid timestamps as compared to current time and nonce life
      */
@@ -69,7 +120,7 @@ class Auth_Sig_Tests extends PHPUnit_Framework_TestCase {
         $little_in_the_past   = $now - ( YOURLS_NONCE_LIFE / 2 );
         $far_in_the_future    = $now + ( YOURLS_NONCE_LIFE * 2 );
         $far_in_the_past      = $now - ( YOURLS_NONCE_LIFE * 2 );
-        
+
         return array(
             array( 0, false ),
             array( $now, true ),
