@@ -5,11 +5,11 @@
  *
  * @group http
  */
- 
-class Misc_HTTP_Tests extends PHPUnit_Framework_TestCase {
+
+class Misc_HTTP_Tests extends PHPUnit\Framework\TestCase {
 
     public function test_get_user_agent() {
-        $this->assertInternalType("string", yourls_get_user_agent());
+        $this->assertIsString(yourls_get_user_agent());
     }
 
     public function test_get_user_agent_empty() {
@@ -19,6 +19,10 @@ class Misc_HTTP_Tests extends PHPUnit_Framework_TestCase {
         $this->assertSame("-", yourls_get_user_agent());
 
         $_SERVER['HTTP_USER_AGENT'] = $copy;
+    }
+
+    public function test_yourls_skip_version_check_is_bool() {
+        $this->assertIsBool(yourls_skip_version_check());
     }
 
 }
