@@ -6,14 +6,14 @@
  * @group l10n
  * @since 0.1
  */
-class Translation_Translation_Tests extends PHPUnit_Framework_TestCase {
-    
-    public static function setUpBeforeClass() {
+class Translation_Translation_Tests extends PHPUnit\Framework\TestCase {
+
+    public static function setUpBeforeClass(): void {
         yourls_load_textdomain( 'test', YOURLS_TESTDATA_DIR . '/pomo/test-fr_FR.mo' );
         yourls_load_textdomain( 'default', YOURLS_TESTDATA_DIR . '/pomo/fr_FR.mo' );
     }
 
-    public static function tearDownAfterClass() {
+    public static function tearDownAfterClass(): void {
         yourls_unload_textdomain( 'test' );
         yourls_unload_textdomain( 'default' );
     }
@@ -26,7 +26,7 @@ class Translation_Translation_Tests extends PHPUnit_Framework_TestCase {
     public function test_translation() {
         $this->assertSame( 'Court (default)' , yourls__( 'Short' ) );
         $this->assertSame( 'Court (test)' ,    yourls__( 'Short', 'test' ) );
-    }    
+    }
 
     /**
      * Check a sample translation - echoed
@@ -36,10 +36,10 @@ class Translation_Translation_Tests extends PHPUnit_Framework_TestCase {
     public function test_translation_echo() {
         $this->expectOutputString( 'Court (default)' );
         yourls_e( 'Short' );
-    }    
+    }
 
     /**
-     * Check an unstranslated string
+     * Check an untranslated string
      *
      * @since 0.1
      */
@@ -82,10 +82,16 @@ class Translation_Translation_Tests extends PHPUnit_Framework_TestCase {
     /**
      * Sprintf'ed with too few arguments - trigger sprintf "too few arguments" error
      *
-     * @expectedException PHPUnit_Framework_Error
      * @since 0.1
      */
     public function test_yourls_s_too_few() {
+        if (PHP_VERSION_ID >= 80000) {
+            $this->expectException(ArgumentCountError::class);
+        } else {
+            $this->expectException(PHPUnit\Framework\Error\Error::class);
+        }
+        $this->expectExceptionMessageMatches('/arguments/');
+
         yourls_s( 'Hello %s you are %s', 'Ozh' );
     }
 
@@ -101,7 +107,7 @@ class Translation_Translation_Tests extends PHPUnit_Framework_TestCase {
 
         // Extra arguments with the last one not being a valid domain: string should be translated
         $this->assertSame( 'Hello Ozh you are nice', yourls_s( 'Hello %s you are %s', 'Ozh', 'nice', 'omg' ) );
-        
+
         // Extra arguments with the last one being a valid domain: string should be translated
         $this->assertSame( 'Bonjour Ozh tu es nice (test)', yourls_s( 'Hello %s you are %s', 'Ozh', 'nice', 'omg', 'test' ) );
     }
@@ -144,7 +150,7 @@ class Translation_Translation_Tests extends PHPUnit_Framework_TestCase {
     public function test_yourls_n() {
         $this->assertSame( '1 truc (default)',   yourls_n( '1 item', '%s items', 1 ) );
         $this->assertSame( '%s trucs (default)', yourls_n( '1 item', '%s items', 2 ) );
-        
+
         $this->assertSame( '1 truc (test)',   yourls_n( '1 item', '%s items', 1, 'test' ) );
         $this->assertSame( '%s trucs (test)', yourls_n( '1 item', '%s items', 2, 'test' ) );
     }
