@@ -5,33 +5,24 @@
  *
  * @group utils
  */
- 
-class NextDecimal_Tests extends PHPUnit_Framework_TestCase {
+
+class NextDecimal_Tests extends PHPUnit\Framework\TestCase {
 
     public function test_get_next_decimal() {
         $id = yourls_get_next_decimal();
-        $this->assertInternalType("int", $id);
-
-        return $id;
+        $this->assertIsInt($id);
     }
 
-	/**
-	 * @depends test_get_next_decimal
-	 */
-    public function test_update_next_decimal_no_arg($id) {
-        $update = yourls_update_next_decimal();
-        $this->assertTrue($update);
+    public function test_update_next_decimal_no_arg() {
+        $id = yourls_get_next_decimal();
+        yourls_update_next_decimal();
         $next = yourls_get_next_decimal();
         $this->assertSame($next, $id + 1);
     }
 
-    /**
-     * @depends test_get_next_decimal
-     */
-    public function test_update_next_decimal_with_arg($id) {
+    public function test_update_next_decimal_with_arg() {
         $rand   = mt_rand(1500,2000);
-        $update = yourls_update_next_decimal($rand);
-        $this->assertTrue($update);
+        yourls_update_next_decimal($rand);
         $next = yourls_get_next_decimal();
         $this->assertSame($next, $rand);
     }
