@@ -2,7 +2,6 @@
 define( 'YOURLS_ADMIN', true );
 define( 'YOURLS_INSTALLING', true );
 require_once( dirname( __DIR__ ).'/includes/load-yourls.php' );
-require_once( YOURLS_INC.'/functions-install.php' );
 
 $error   = array();
 $warning = array();
@@ -57,7 +56,7 @@ yourls_html_head( 'install', yourls__( 'Install YOURLS' ) );
 <div id="login">
 	<form method="post" action="?"><?php // reset any QUERY parameters ?>
 		<p>
-			<img src="<?php yourls_site_url(); ?>/images/yourls-logo.png" alt="YOURLS" title="YOURLS" />
+			<img src="<?php yourls_site_url(); ?>/images/yourls-logo.svg" id="yourls-logo" alt="YOURLS" title="YOURLS" />
 		</p>
 		<?php
 			// Print errors, warnings and success messages
