@@ -6,7 +6,7 @@ $auth = yourls_is_valid_user();
 
 if( $auth !== true ) {
 
-	// API mode, 
+	// API mode,
 	if ( yourls_is_API() ) {
 		$format = ( isset($_REQUEST['format']) ? $_REQUEST['format'] : 'xml' );
 		$callback = ( isset($_REQUEST['callback']) ? $_REQUEST['callback'] : '' );
@@ -21,7 +21,7 @@ if( $auth !== true ) {
 	} else {
 		yourls_login_screen( $auth );
 	}
-	
+
 	die();
 }
 
@@ -33,7 +33,7 @@ yourls_do_action( 'auth_successful' );
  *
  * TODO: Remove this once real user management is implemented
  */
- 
+
 // Did we just fail at encrypting passwords ?
 if ( isset( $_GET['dismiss'] ) && $_GET['dismiss'] == 'hasherror' ) {
 	yourls_update_option( 'defer_hashing_error', time() + 86400 * 7 ); // now + 1 week
@@ -41,23 +41,23 @@ if ( isset( $_GET['dismiss'] ) && $_GET['dismiss'] == 'hasherror' ) {
 } else {
 
 	// Encrypt passwords that are clear text
-	if ( !defined( 'YOURLS_NO_HASH_PASSWORD' ) && yourls_has_cleartext_passwords() ) {
-		$hash = yourls_hash_passwords_now( YOURLS_CONFIGFILE );
-		if ( $hash === true ) {
-			// Hashing succesful. Remove flag from DB if any.
-			if( yourls_get_option( 'defer_hashing_error' ) )
-				yourls_delete_option( 'defer_hashing_error' );
-		} else {
-			// It failed, display message for first time or if last time was a week ago
-			if ( time() > yourls_get_option( 'defer_hashing_error' ) or !yourls_get_option( 'defer_hashing_error' ) ) {
-				$message  = yourls_s( 'Could not auto-encrypt passwords. Error was: "%s".', $hash );
-				$message .= ' ';
-				$message .= yourls_s( '<a href="%s">Get help</a>.', 'http://yourls.org/userpassword' );
-				$message .= '</p><p>';
-				$message .= yourls_s( '<a href="%s">Click here</a> to dismiss this message for one week.', '?dismiss=hasherror' );
-				
-				yourls_add_notice( $message );
-			}
-		}
+	if ( yourls_maybe_hash_passwords() ) {
+        $hash = yourls_hash_passwords_now( YOURLS_CONFIGFILE );
+        if ( $hash === true ) {
+            // Hashing successful. Remove flag from DB if any.
+            if( yourls_get_option( 'defer_hashing_error' ) )
+                yourls_delete_option( 'defer_hashing_error' );
+        } else {
+            // It failed, display message for first time or if last time was a week ago
+            if ( time() > yourls_get_option( 'defer_hashing_error' ) or !yourls_get_option( 'defer_hashing_error' ) ) {
+                $message  = yourls_s( 'Could not auto-encrypt passwords. Error was: "%s".', $hash );
+                $message .= ' ';
+                $message .= yourls_s( '<a href="%s">Get help</a>.', 'http://yourls.org/userpassword' );
+                $message .= '</p><p>';
+                $message .= yourls_s( '<a href="%s">Click here</a> to dismiss this message for one week.', '?dismiss=hasherror' );
+
+                yourls_add_notice( $message );
+            }
+        }
 	}
 }
