@@ -163,8 +163,12 @@ function yourls_api_output( $mode, $output, $send_headers = true, $echo = true )
 /**
  * Return array for API stat requests
  *
+ * @param string $filter  either "top", "bottom" , "rand" or "last"
+ * @param int    $limit   maximum number of links to return
+ * @param int    $start   offset
+ * @return array
  */
-function yourls_api_stats( $filter = 'top', $limit = 10, $start = 0 ) {
+function yourls_api_stats($filter = 'top', $limit = 10, $start = 0 ) {
 	$return = yourls_get_stats( $filter, $limit, $start );
 	$return['simple']  = 'Need either XML or JSON format for stats';
 	$return['message'] = 'success';
@@ -174,6 +178,7 @@ function yourls_api_stats( $filter = 'top', $limit = 10, $start = 0 ) {
 /**
  * Return array for counts of shorturls and clicks
  *
+ * @return array
  */
 function yourls_api_db_stats() {
 	$return = array(
@@ -189,12 +194,14 @@ function yourls_api_db_stats() {
 /**
  * Return array for API stat requests
  *
+ * @param string $shorturl  Short URL to check
+ * @return array
  */
 function yourls_api_url_stats( $shorturl ) {
-	$keyword = str_replace( YOURLS_SITE . '/' , '', $shorturl ); // accept either 'http://ozh.in/abc' or 'abc'
-	$keyword = yourls_sanitize_string( $keyword );
+	$keyword = str_replace( yourls_get_yourls_site() . '/' , '', $shorturl ); // accept either 'http://ozh.in/abc' or 'abc'
+	$keyword = yourls_sanitize_keyword( $keyword );
 
-	$return = yourls_get_link_stats( $keyword );
+	$return = yourls_get_keyword_stats( $keyword );
 	$return['simple']  = 'Need either XML or JSON format for stats';
 	return yourls_apply_filter( 'api_url_stats', $return, $shorturl );
 }
@@ -202,17 +209,19 @@ function yourls_api_url_stats( $shorturl ) {
 /**
  * Expand short url to long url
  *
+ * @param string $shorturl  Short URL to expand
+ * @return array
  */
 function yourls_api_expand( $shorturl ) {
-	$keyword = str_replace( YOURLS_SITE . '/' , '', $shorturl ); // accept either 'http://ozh.in/abc' or 'abc'
-	$keyword = yourls_sanitize_string( $keyword );
+	$keyword = str_replace( yourls_get_yourls_site() . '/' , '', $shorturl ); // accept either 'http://ozh.in/abc' or 'abc'
+	$keyword = yourls_sanitize_keyword( $keyword );
 
 	$longurl = yourls_get_keyword_longurl( $keyword );
 
 	if( $longurl ) {
 		$return = array(
 			'keyword'   => $keyword,
-			'shorturl'  => YOURLS_SITE . "/$keyword",
+			'shorturl'  => yourls_link($keyword),
 			'longurl'   => $longurl,
             'title'     => yourls_get_keyword_title( $keyword ),
 			'simple'    => $longurl,
