@@ -1,4 +1,5 @@
 <?php
+
 /**
  * Functions that relate to HTTP requests
  *
@@ -14,6 +15,8 @@
  * @since 1.7
  */
 
+use WpOrg\Requests\Requests;
+
 /**
  * Perform a GET request, return response object or error string message
  *
@@ -21,6 +24,10 @@
  *
  * @since 1.7
  * @see yourls_http_request
+ * @param string $url     URL to request
+ * @param array $headers  HTTP headers to send
+ * @param array $data     GET data
+ * @param array $options  Options to pass to Requests
  * @return mixed Response object, or error string
  */
 function yourls_http_get( $url, $headers = array(), $data = array(), $options = array() ) {
@@ -32,6 +39,10 @@ function yourls_http_get( $url, $headers = array(), $data = array(), $options = 
  *
  * @since 1.7
  * @see yourls_http_request
+ * @param string $url     URL to request
+ * @param array $headers  HTTP headers to send
+ * @param array $data     GET data
+ * @param array $options  Options to pass to Requests
  * @return mixed String (page body) or null if error
  */
 function yourls_http_get_body( $url, $headers = array(), $data = array(), $options = array() ) {
@@ -46,6 +57,10 @@ function yourls_http_get_body( $url, $headers = array(), $data = array(), $optio
  *
  * @since 1.7
  * @see yourls_http_request
+ * @param string $url     URL to request
+ * @param array $headers  HTTP headers to send
+ * @param array $data     POST data
+ * @param array $options  Options to pass to Requests
  * @return mixed Response object, or error string
  */
 function yourls_http_post( $url, $headers = array(), $data = array(), $options = array() ) {
@@ -59,6 +74,10 @@ function yourls_http_post( $url, $headers = array(), $data = array(), $options =
  *
  * @since 1.7
  * @see yourls_http_request
+ * @param string $url     URL to request
+ * @param array $headers  HTTP headers to send
+ * @param array $data     POST data
+ * @param array $options  Options to pass to Requests
  * @return mixed String (page body) or null if error
  */
 function yourls_http_post_body( $url, $headers = array(), $data = array(), $options = array() ) {
@@ -152,7 +171,7 @@ function yourls_send_through_proxy( $url ) {
 		return true;
 
 	// Self and loopback URLs are considered local (':' is parse_url() host on '::1')
-	$home = parse_url( YOURLS_SITE );
+	$home = parse_url( yourls_get_yourls_site() );
 	$local = array( 'localhost', '127.0.0.1', '127.1', '[::1]', ':', $home['host'] );
 
 	if( in_array( $check['host'], $local ) )
@@ -197,7 +216,7 @@ function yourls_send_through_proxy( $url ) {
  * @param array $headers Extra headers to send with the request
  * @param array $data Data to send either as a query string for GET requests, or in the body for POST requests
  * @param array $options Options for the request (see /includes/Requests/Requests.php:request())
- * @return object Requests_Response object
+ * @return object WpOrg\Requests\Response object
  */
 function yourls_http_request( $type, $url, $headers, $data, $options ) {
 
@@ -206,33 +225,26 @@ function yourls_http_request( $type, $url, $headers, $data, $options ) {
 	if ( null !== $pre )
 		return $pre;
 
-	yourls_http_load_library();
-
 	$options = array_merge( yourls_http_default_options(), $options );
 
 	if( yourls_http_get_proxy() && !yourls_send_through_proxy( $url ) ) {
 		unset( $options['proxy'] );
 	}
 
+    // filter everything
+    $type    = yourls_apply_filter('http_request_type', $type);
+    $url     = yourls_apply_filter('http_request_url', $url);
+    $headers = yourls_apply_filter('http_request_headers', $headers);
+    $data    = yourls_apply_filter('http_request_data', $data);
+    $options = yourls_apply_filter('http_request_options', $options);
+
 	try {
 		$result = Requests::request( $url, $headers, $data, $type, $options );
-	} catch( Requests_Exception $e ) {
+	} catch( \WpOrg\Requests\Exception $e ) {
 		$result = yourls_debug_log( $e->getMessage() . ' (' . $type . ' on ' . $url . ')' );
 	};
 
 	return $result;
-}
-
-/**
- * Include Requests library if need be
- *
- * This is to avoid include()-ing all the Requests files on every YOURLS instance
- * disregarding whether needed or not.
- *
- * @since 1.7
- */
-function yourls_http_load_library() {
-    Requests::register_autoloader();
 }
 
 /**
@@ -242,7 +254,7 @@ function yourls_http_load_library() {
  * @return string UA string
  */
 function yourls_http_user_agent() {
-	return yourls_apply_filter( 'http_user_agent', 'YOURLS v'.YOURLS_VERSION.' +http://yourls.org/ (running on '.YOURLS_SITE.')' );
+	return yourls_apply_filter( 'http_user_agent', 'YOURLS v'.YOURLS_VERSION.' +http://yourls.org/ (running on '.yourls_get_yourls_site().')' );
 }
 
 /**
@@ -262,7 +274,7 @@ function yourls_http_user_agent() {
  */
 function yourls_check_core_version() {
 
-	global $ydb, $yourls_user_passwords;
+	global $yourls_user_passwords;
 
 	$checks = yourls_get_option( 'core_version_checks' );
 
@@ -279,22 +291,21 @@ function yourls_check_core_version() {
 		$checks->version_checked = YOURLS_VERSION;
 	}
 
-	// Config file location ('u' for '/user' or 'i' for '/includes')
-	$conf_loc = str_replace( YOURLS_ABSPATH, '', YOURLS_CONFIGFILE );
-	$conf_loc = str_replace( '/config.php', '', $conf_loc );
-	$conf_loc = ( $conf_loc == '/user' ? 'u' : 'i' );
+	// Total number of links and clicks
+    list( $total_urls, $total_clicks ) = array_values(yourls_get_db_stats());
 
 	// The collection of stuff to report
 	$stuff = array(
 		// Globally uniquish site identifier
+        // This uses const YOURLS_SITE and not yourls_get_yourls_site() to prevent creating another id for an already known install
 		'md5'                => md5( YOURLS_SITE . YOURLS_ABSPATH ),
 
 		// Install information
 		'failed_attempts'    => $checks->failed_attempts,
-		'yourls_site'        => defined( 'YOURLS_SITE' ) ? YOURLS_SITE : 'unknown',
+		'yourls_site'        => defined( 'YOURLS_SITE' ) ? yourls_get_yourls_site() : 'unknown',
 		'yourls_version'     => defined( 'YOURLS_VERSION' ) ? YOURLS_VERSION : 'unknown',
 		'php_version'        => PHP_VERSION,
-		'mysql_version'      => $ydb->mysql_version(),
+		'mysql_version'      => yourls_get_db()->mysql_version(),
 		'locale'             => yourls_get_locale(),
 
 		// custom DB driver if any, and useful common PHP extensions
@@ -305,21 +316,25 @@ function yourls_check_core_version() {
 		'ext_curl'           => extension_loaded( 'curl' )    ? 1 : 0,
 
 		// Config information
-		'num_users'          => count( $yourls_user_passwords ),
-		'config_location'    => $conf_loc,
 		'yourls_private'     => defined( 'YOURLS_PRIVATE' ) && YOURLS_PRIVATE ? 1 : 0,
 		'yourls_unique'      => defined( 'YOURLS_UNIQUE_URLS' ) && YOURLS_UNIQUE_URLS ? 1 : 0,
 		'yourls_url_convert' => defined( 'YOURLS_URL_CONVERT' ) ? YOURLS_URL_CONVERT : 'unknown',
-		'num_active_plugins' => yourls_has_active_plugins(),
-		'num_pages'          => defined( 'YOURLS_PAGEDIR' ) ? count( (array) glob( YOURLS_PAGEDIR .'/*.php') ) : 0,
+
+        // Usage information
+        'num_users'          => count( $yourls_user_passwords ),
+        'num_active_plugins' => yourls_has_active_plugins(),
+        'num_pages'          => defined( 'YOURLS_PAGEDIR' ) ? count( (array) glob( YOURLS_PAGEDIR .'/*.php') ) : 0,
+        'num_links'          => $total_urls,
+        'num_clicks'         => $total_clicks,
 	);
 
 	$stuff = yourls_apply_filter( 'version_check_stuff', $stuff );
 
 	// Send it in
-	$url = 'http://api.yourls.org/core/version/1.0/';
-    if( yourls_can_http_over_ssl() )
-        $url = yourls_set_url_scheme( $url, 'https' );
+	$url = 'http://api.yourls.org/core/version/1.1/';
+    if( yourls_can_http_over_ssl() ) {
+        $url = yourls_set_url_scheme($url, 'https');
+    }
 	$req = yourls_http_post( $url, array(), $stuff );
 
 	$checks->last_attempt = time();
@@ -329,13 +344,16 @@ function yourls_check_core_version() {
 	if( is_string( $req ) or !$req->success ) {
 		$checks->failed_attempts = $checks->failed_attempts + 1;
 		yourls_update_option( 'core_version_checks', $checks );
+		if( is_string($req) ) {
+            yourls_debug_log('Version check failed: ' . $req);
+        }
 		return false;
 	}
 
 	// Parse response
 	$json = json_decode( trim( $req->body ) );
 
-	if( isset( $json->latest ) && isset( $json->zipurl ) ) {
+	if( yourls_validate_core_version_response($json) ) {
 		// All went OK - mark this down
 		$checks->failed_attempts = 0;
 		$checks->last_result     = $json;
@@ -349,6 +367,83 @@ function yourls_check_core_version() {
 }
 
 /**
+ *  Make sure response from api.yourls.org is valid
+ *
+ *  1) we should get a json object with two following properties:
+ *    'latest' => a string representing a YOURLS version number, eg '1.2.3'
+ *    'zipurl' => a string for a zip package URL, from github, eg 'https://api.github.com/repos/YOURLS/YOURLS/zipball/1.2.3'
+ *  2) 'latest' and version extracted from 'zipurl' should match
+ *  3) the object should not contain any other key
+ *
+ *  @since 1.7.7
+ *  @param object $json  JSON object to check
+ *  @return bool   true if seems legit, false otherwise
+ */
+function yourls_validate_core_version_response($json) {
+    return (
+        yourls_validate_core_version_response_keys($json)
+     && $json->latest === yourls_sanitize_version($json->latest)
+     && $json->zipurl === yourls_sanitize_url($json->zipurl)
+     && $json->latest === yourls_get_version_from_zipball_url($json->zipurl)
+     && yourls_is_valid_github_repo_url($json->zipurl)
+    );
+}
+
+/**
+ * Get version number from Github zipball URL (last part of URL, really)
+ *
+ * @since 1.8.3
+ * @param string $zipurl eg 'https://api.github.com/repos/YOURLS/YOURLS/zipball/1.2.3'
+ * @return string
+ */
+function yourls_get_version_from_zipball_url($zipurl) {
+    $version = '';
+    $parts = explode('/', parse_url(yourls_sanitize_url($zipurl), PHP_URL_PATH) ?? '');
+    // expect at least 1 slash in path, return last part
+    if( count($parts) > 1 ) {
+        $version = end($parts);
+    }
+    return $version;
+}
+
+/**
+ * Check if URL is from YOURLS/YOURLS repo on github
+ *
+ * @since 1.8.3
+ * @param string $url  URL to check
+ * @return bool
+ */
+function yourls_is_valid_github_repo_url($url) {
+    $url = yourls_sanitize_url($url);
+    return (
+        join('.',array_slice(explode('.', parse_url($url, PHP_URL_HOST) ?? ''), -2, 2)) === 'github.com'
+            // explodes on '.' (['api','github','com']) and keeps the last two elements
+            // to make sure domain is either github.com or one of its subdomain (api.github.com for instance)
+            // TODO: keep an eye on Github API to make sure it doesn't change some day to another domain (githubapi.com, ...)
+        && substr( parse_url($url, PHP_URL_PATH), 0, 21 ) === '/repos/YOURLS/YOURLS/'
+            // make sure path starts with '/repos/YOURLS/YOURLS/'
+    );
+}
+
+/**
+ * Check if object has only expected keys 'latest' and 'zipurl' containing strings
+ *
+ * @since 1.8.3
+ * @param object $json
+ * @return bool
+ */
+function yourls_validate_core_version_response_keys($json) {
+    $keys = array('latest', 'zipurl');
+    return (
+        count(array_diff(array_keys((array)$json), $keys)) === 0
+        && isset($json->latest)
+        && isset($json->zipurl)
+        && is_string($json->latest)
+        && is_string($json->zipurl)
+    );
+}
+
+/**
  * Determine if we want to check for a newer YOURLS version (and check if applicable)
  *
  * Currently checks are performed every 24h and only when someone is visiting an admin page.
@@ -358,17 +453,19 @@ function yourls_check_core_version() {
  * @return bool true if a check was needed and successfully performed, false otherwise
  */
 function yourls_maybe_check_core_version() {
+    // Allow plugins to short-circuit the whole function
+    $pre = yourls_apply_filter('shunt_maybe_check_core_version', null);
+    if (null !== $pre) {
+        return $pre;
+    }
 
-	// Allow plugins to short-circuit the whole function
-	$pre = yourls_apply_filter( 'shunt_maybe_check_core_version', null );
-	if ( null !== $pre )
-		return $pre;
+    if (yourls_skip_version_check()) {
+        return false;
+    }
 
-	if( defined( 'YOURLS_NO_VERSION_CHECK' ) && YOURLS_NO_VERSION_CHECK )
-		return false;
-
-	if( !yourls_is_admin() )
-		return false;
+    if (!yourls_is_admin()) {
+        return false;
+    }
 
 	$checks = yourls_get_option( 'core_version_checks' );
 
@@ -400,6 +497,16 @@ function yourls_maybe_check_core_version() {
 }
 
 /**
+ * Check if user setting for skipping version check is set
+ *
+ * @since 1.8.2
+ * @return bool
+ */
+function yourls_skip_version_check() {
+    return yourls_apply_filter('skip_version_check', defined('YOURLS_NO_VERSION_CHECK') && YOURLS_NO_VERSION_CHECK);
+}
+
+/**
  * Check if server can perform HTTPS requests, return bool
  *
  * @since 1.7.1
@@ -419,4 +526,3 @@ function yourls_can_http_over_ssl() {
 
     return ( $ssl_curl OR $ssl_socket );
 }
-
