@@ -13,6 +13,7 @@ function yourls_check_PDO() {
 /**
  * Check if server has MySQL 5.0+
  *
+ * @return bool
  */
 function yourls_check_database_version() {
     return ( version_compare( '5.0', yourls_get_database_version() ) <= 0 );
@@ -27,25 +28,29 @@ function yourls_check_database_version() {
 function yourls_get_database_version() {
 	// Allow plugins to short-circuit the whole function
 	$pre = yourls_apply_filter( 'shunt_get_database_version', false );
-	if ( false !== $pre )
+	if ( false !== $pre ) {
 		return $pre;
+    }
 
-	global $ydb;
-
-	return yourls_sanitize_version($ydb->mysql_version());
+	return yourls_sanitize_version(yourls_get_db()->mysql_version());
 }
 
 /**
- * Check if PHP > 5.6
+ * Check if PHP > 7.2
  *
+ * As of 1.8 we advertise YOURLS as being 7.4+ but it should work on 7.2 (although untested)
+ * so we don't want to strictly enforce a limitation that may not be necessary.
+ *
+ * @return bool
  */
 function yourls_check_php_version() {
-    return version_compare( PHP_VERSION, '5.6.0', '>=' );
+    return version_compare( PHP_VERSION, '7.2.0', '>=' );
 }
 
 /**
  * Check if server is an Apache
  *
+ * @return bool
  */
 function yourls_is_apache() {
 	if( !array_key_exists( 'SERVER_SOFTWARE', $_SERVER ) )
@@ -59,6 +64,7 @@ function yourls_is_apache() {
 /**
  * Check if server is running IIS
  *
+ * @return bool
  */
 function yourls_is_iis() {
 	return ( array_key_exists( 'SERVER_SOFTWARE', $_SERVER ) ? ( strpos( $_SERVER['SERVER_SOFTWARE'], 'IIS' ) !== false ) : false );
@@ -68,9 +74,10 @@ function yourls_is_iis() {
 /**
  * Create .htaccess or web.config. Returns boolean
  *
+ * @return bool
  */
 function yourls_create_htaccess() {
-	$host = parse_url( YOURLS_SITE );
+	$host = parse_url( yourls_get_yourls_site() );
 	$path = ( isset( $host['path'] ) ? $host['path'] : '' );
 
 	if ( yourls_is_iis() ) {
@@ -200,7 +207,7 @@ function yourls_create_sql_tables() {
         return $pre;
     }
 
-	global $ydb;
+	$ydb = yourls_get_db();
 
 	$error_msg = array();
 	$success_msg = array();
@@ -208,39 +215,39 @@ function yourls_create_sql_tables() {
 	// Create Table Query
 	$create_tables = array();
 	$create_tables[YOURLS_DB_TABLE_URL] =
-		'CREATE TABLE IF NOT EXISTS `'.YOURLS_DB_TABLE_URL.'` ('.
-		'`keyword` varchar(200) BINARY NOT NULL,'.
-		'`url` text BINARY NOT NULL,'.
-		'`title` text CHARACTER SET utf8,'.
-		'`timestamp` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,'.
-		'`ip` VARCHAR(41) NOT NULL,'.
-		'`clicks` INT(10) UNSIGNED NOT NULL,'.
-		' PRIMARY KEY  (`keyword`),'.
-		' KEY `timestamp` (`timestamp`),'.
-		' KEY `ip` (`ip`)'.
-		');';
+        'CREATE TABLE IF NOT EXISTS `'.YOURLS_DB_TABLE_URL.'` ('.
+         '`keyword` varchar(100) CHARACTER SET utf8mb4 COLLATE utf8mb4_bin NOT NULL DEFAULT "",'.
+         '`url` text CHARACTER SET utf8mb4 COLLATE utf8mb4_bin NOT NULL,'.
+         '`title` text COLLATE utf8mb4_unicode_ci DEFAULT NULL,'.
+         '`timestamp` timestamp NOT NULL DEFAULT current_timestamp(),'.
+         '`ip` varchar(41) COLLATE utf8mb4_unicode_ci NOT NULL,'.
+         '`clicks` int(10) unsigned NOT NULL,'.
+         'PRIMARY KEY (`keyword`),'.
+         'KEY `ip` (`ip`),'.
+         'KEY `timestamp` (`timestamp`)'.
+        ') DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin;';
 
 	$create_tables[YOURLS_DB_TABLE_OPTIONS] =
 		'CREATE TABLE IF NOT EXISTS `'.YOURLS_DB_TABLE_OPTIONS.'` ('.
 		'`option_id` bigint(20) unsigned NOT NULL auto_increment,'.
-		'`option_name` varchar(64) NOT NULL default "",'.
-		'`option_value` longtext NOT NULL,'.
+		'`option_name` varchar(64) COLLATE utf8mb4_unicode_ci NOT NULL default "",'.
+		'`option_value` longtext COLLATE utf8mb4_unicode_ci NOT NULL,'.
 		'PRIMARY KEY  (`option_id`,`option_name`),'.
 		'KEY `option_name` (`option_name`)'.
-		') AUTO_INCREMENT=1 ;';
+		') AUTO_INCREMENT=1 DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;';
 
 	$create_tables[YOURLS_DB_TABLE_LOG] =
 		'CREATE TABLE IF NOT EXISTS `'.YOURLS_DB_TABLE_LOG.'` ('.
 		'`click_id` int(11) NOT NULL auto_increment,'.
 		'`click_time` datetime NOT NULL,'.
-		'`shorturl` varchar(200) BINARY NOT NULL,'.
+		'`shorturl` varchar(100) BINARY NOT NULL,'.
 		'`referrer` varchar(200) NOT NULL,'.
 		'`user_agent` varchar(255) NOT NULL,'.
 		'`ip_address` varchar(41) NOT NULL,'.
 		'`country_code` char(2) NOT NULL,'.
 		'PRIMARY KEY  (`click_id`),'.
 		'KEY `shorturl` (`shorturl`)'.
-		') AUTO_INCREMENT=1 ;';
+		') AUTO_INCREMENT=1 DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;';
 
 
 	$create_table_count = 0;
@@ -303,9 +310,9 @@ function yourls_initialize_options() {
  * @return bool
  */
 function yourls_insert_sample_links() {
-	$link1 = yourls_add_new_link( 'http://blog.yourls.org/', 'yourlsblog', 'YOURLS\' Blog' );
-	$link2 = yourls_add_new_link( 'http://yourls.org/',      'yourls',     'YOURLS: Your Own URL Shortener' );
-	$link3 = yourls_add_new_link( 'http://ozh.org/',         'ozh',        'ozh.org' );
+	$link1 = yourls_add_new_link( 'https://blog.yourls.org/', 'yourlsblog', 'YOURLS\' Blog' );
+	$link2 = yourls_add_new_link( 'https://yourls.org/',      'yourls',     'YOURLS: Your Own URL Shortener' );
+	$link3 = yourls_add_new_link( 'https://ozh.org/',         'ozh',        'ozh.org' );
 	return ( bool ) (
 		  $link1['status'] == 'success'
 		& $link2['status'] == 'success'
@@ -317,6 +324,8 @@ function yourls_insert_sample_links() {
 /**
  * Toggle maintenance mode. Inspired from WP. Returns true for success, false otherwise
  *
+ * @param bool $maintenance  True to enable, false to disable
+ * @return bool              True on success, false on failure
  */
 function yourls_maintenance_mode( $maintenance = true ) {
 
@@ -340,4 +349,3 @@ function yourls_maintenance_mode( $maintenance = true ) {
 		return @unlink($file);
 	}
 }
-
