@@ -2,11 +2,19 @@
 /**
  * YOURLS version
  *
+ * Must be one of the following :
+ *    MAJOR.MINOR (eg 1.8)
+ *    MAJOR.MINOR.PATCH (1.8.1)
+ *    MAJOR.MINOR-SOMETHING (1.8-dev)
+ *    MAJOR.MINOR.PATCH-SOMETHING (1.8.1-donotuse)
+ *
  */
-define( 'YOURLS_VERSION', '1.7.7' );
+define( 'YOURLS_VERSION', '1.9.3-dev' );
 
 /**
  * YOURLS DB version. Increments when changes are made to the DB schema, to trigger a DB update
  *
+ * Must be a string of an integer.
+ *
  */
-define( 'YOURLS_DB_VERSION', '482' );
+define( 'YOURLS_DB_VERSION', '506' );
