@@ -9,11 +9,11 @@ $(document).ready(function(){
 function update_share() {
 	var text = encodeURIComponent( $('#tweet_body').val() );
 	var url = encodeURIComponent( $('#copylink').val() );
-	var tw = 'https://twitter.com/intent/tweet?status='+text;
+	var tw = 'https://twitter.com/intent/tweet?text='+text;
 	var fb = 'https://www.facebook.com/share.php?u='+url ;
 	$('#share_tw').attr('href', tw);
 	$('#share_fb').attr('href', fb);
-	
+
 	var charcount = parseInt(280 - $('#tweet_body').val().length);
 	$('#charcount')
 		.toggleClass("negative", charcount < 0)
@@ -36,12 +36,12 @@ function share(dest) {
 }
 
 function init_clipboard() {
-    var clipboard = new Clipboard('#copylink', {
+    var clipboard = new ClipboardJS('#copylink', {
         text: function (trigger) {
             return $(trigger).val();
         }
     });
-    
+
     clipboard.on('success', function () {
         $('#copylink').select();
         html_pulse('#copybox h2, #copybox h3', 'Copied!');
