@@ -1,6 +1,6 @@
 <?php
 /**
- * YOURLS Config for local unit tests. Copy this file to yourls-test-config.php
+ * YOURLS Config for local unit tests. Copy this file to yourls-tests-config.php
  */
 
 /*** YOURLS code base you want to test */
@@ -28,7 +28,6 @@ define('YOURLS_URL_CONVERT',  62);
 define('YOURLS_DB_PREFIX',  'yourls_');
 define('YOURLS_FLOOD_DELAY_SECONDS',  0);
 define('YOURLS_FLOOD_IP_WHITELIST',  '');
-define('YOURLS_NO_HASH_PASSWORD',  true); // prevents rewriting config.php with encrypted passwords
 define('YOURLS_LANG',  'fr_FR'); // locale of a sample translation file in the data dir
 define('YOURLS_DEBUG', true);
 
@@ -37,11 +36,13 @@ $yourls_reserved_URL = array(
 );
 
 $yourls_user_passwords = array(
-	'yourls'  => 'travis-ci-test',
+	'yourls'  => 'secret-ci-test',
 	'clear'   => 'somepassword',
 	'md5'     => 'md5:12373:e52e4488f79a740bd341f229e3c163c8',                          // password: '3cd6944201fa7bbc5e0fe852e36b1096' with md5 and salt
-	'phpass'  => 'phpass:!2a!08!T1ptMlBSxu7g3odpbUXgd.9wbKvg8k7cJt.HbwSqUNrlLPudWnf/6', // password: '3cd6944201fa7bbc5e0fe852e36b1096' with PHPass
-	'phpass2' => 'phpass:$2a$08$gt2bnpfUyuCX3hrp0RPOieFR1RwBnLsMzpq/NvPXwCdV3LqI3RGYi', // password: also '3cd6944201fa7bbc5e0fe852e36b1096' but without YOURLS internal char substitution
+    'phpass'  => 'phpass:!2a!08!T1ptMlBSxu7g3odpbUXgd.9wbKvg8k7cJt.HbwSqUNrlLPudWnf/6', // password: '3cd6944201fa7bbc5e0fe852e36b1096' with old PHPass library
+    'phpass2' => 'phpass:$2a$08$gt2bnpfUyuCX3hrp0RPOieFR1RwBnLsMzpq/NvPXwCdV3LqI3RGYi', // password: also '3cd6944201fa7bbc5e0fe852e36b1096' with old PHPass lib but without YOURLS internal char substitution
+    'phpass3' => 'phpass:!2y!10!.FjK.vQR0JVivkMwckiiIesFUFhtMxX/f9pes.i/ccp/W0IuUSxPW', // password: also '3cd6944201fa7bbc5e0fe852e36b1096' hashed with password_hash
+    'phpass4' => 'phpass:$2y$10$KPP/sv7pv0JL2GwcixNBfuXRPElC4KxQUgetqBfCboB.q30yKwKG6', // password: also '3cd6944201fa7bbc5e0fe852e36b1096' hashed with password_hash but without YOURLS internal char substitution
     '1994'    => '@$*',
     'special' => 'lol .\+*?[^]$(){}=!<>|:-/',
     'quote1'  => '"ahah"',
