@@ -60,7 +60,7 @@ function ozh_random_shorturl_settings_page() {
         ozh_random_shorturl_settings_update();
     }
 
-    $random_length = yourls_get_option('random_shorturls_length');
+    $random_length = yourls_get_option('random_shorturls_length', 5);
     $nonce = yourls_create_nonce( 'random_shorturl_settings' );
 
     echo <<<HTML
