@@ -2,8 +2,8 @@
 /*
 Plugin Name: YOURLS Toolbar
 Plugin URI: http://yourls.org/
-Description: Add a social toolbar to your redirected short URLs. Fork this plugin if you want to make your own toolbar.
-Version: 1.0
+Description: Add a toolbar to frame your short URLs. Fork this plugin if you want to make your own toolbar.
+Version: 1.1
 Author: Ozh
 Author URI: http://ozh.org/
 Disclaimer: Toolbars ruin the user experience. Be warned.
@@ -28,15 +28,17 @@ function ozh_toolbar_add( $args ) {
 yourls_add_action( 'pre_redirect', 'ozh_toolbar_do' );
 function ozh_toolbar_do( $args ) {
 	global $ozh_toolbar;
-	
+
 	// Does this redirection need a toolbar?
-	if( !$ozh_toolbar['do'] )
+	if( !$ozh_toolbar['do'] ) {
 		return;
+    }
 
 	// Do we have a cookie stating the user doesn't want a toolbar?
-	if( isset( $_COOKIE['yourls_no_toolbar'] ) && $_COOKIE['yourls_no_toolbar'] == 1 )
+	if( isset( $_COOKIE['yourls_no_toolbar'] ) && $_COOKIE['yourls_no_toolbar'] == 1 ) {
 		return;
-	
+    }
+
 	// Get URL and page title
 	$url = $args[0];
 	$pagetitle = yourls_get_keyword_title( $ozh_toolbar['keyword'] );
@@ -47,11 +49,10 @@ function ozh_toolbar_do( $args ) {
 		yourls_edit_link_title( $ozh_toolbar['keyword'], $pagetitle );
 	}
 	$_pagetitle = htmlentities( yourls_get_remote_title( $url ) );
-	
+
 	$www = YOURLS_SITE;
 	$ver = YOURLS_VERSION;
-	$md5 = md5( $url );
-	$sql = yourls_get_num_queries();
+    $favicon = yourls_get_yourls_favicon_url(false);
 
 	// When was the link created (in days)
 	$diff = abs( time() - strtotime( yourls_get_keyword_timestamp( $ozh_toolbar['keyword'] ) ) );
@@ -61,11 +62,11 @@ function ozh_toolbar_do( $args ) {
 	} else {
 		$created = $days . ' ' . yourls_n( 'day', 'days', $days ) . ' ago';
 	}
-	
+
 	// How many hits on the page
 	$hits = 1 + yourls_get_keyword_clicks( $ozh_toolbar['keyword'] );
 	$hits = $hits . ' ' . yourls_n( 'view', 'views', $hits );
-	
+
 	// Plugin URL (no URL is hardcoded)
 	$pluginurl = YOURLS_PLUGINURL . '/'.yourls_plugin_basename( __DIR__ );
 
@@ -74,9 +75,8 @@ function ozh_toolbar_do( $args ) {
 <html>
 <head>
 	<title>$pagetitle &mdash; YOURLS</title>
-	<link rel="icon" type="image/gif" href="$www/images/favicon.gif" />
+	<link rel="icon" type="image/gif" href="$favicon" />
 	<meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
-	<meta http-equiv="X-UA-Compatible" content="IE-9"/>
 	<meta name="generator" content="YOURLS v$ver" />
 	<meta name="ROBOTS" content="NOINDEX, FOLLOW" />
 	<link rel="stylesheet" href="$pluginurl/css/toolbar.css" type="text/css" media="all" />
@@ -85,42 +85,20 @@ function ozh_toolbar_do( $args ) {
 <div id="yourls-bar">
 	<div id="yourls-about">
 		Short link powered by <a href="http://yourls.org/">YOURLS</a> and created $created. $hits.
-		<!-- $sql queries -->
-	</div>
-	
-	<div id="yourls-delicious">
-	<img src="http://static.delicious.com/img/delicious.small.gif" height="10" width="10" alt="Delicious" />
-	<a id="yourls-delicious-link" title="Bookmark on delicious" href="http://delicious.com/save" onclick="window.open('http://delicious.com/save?v=5&noui&jump=close&url='+encodeURIComponent(location.href)+'&title='+encodeURIComponent(document.title), 'delicious','toolbar=no,width=550,height=550'); return false;"> Bookmark on Delicious</a>
 	</div>
 
-	<script type="text/javascript" id="topsy_global_settings">
-	var topsy_theme = "light-blue";
-	var topsy_nick = " ";
-	var topsy_style = "small";
-	var topsy_order = "count,retweet,badge";
-	</script>
-	<div id="yourls-topsy" class="topsy_widget_data">
-		<!--{
-		        "url": "$www/{$ozh_toolbar['keyword']}",
-		        "title": "$_pagetitle",
-		}-->
-	</div>
-	
 	<div id="yourls-selfclose">
 		<a id="yourls-once" href="$url" title="Close this toolbar">close</a>
 		<a id="yourls-always" href="$url" title="Never show me this toolbar again">close</a>
-		
 	</div>
 </div>
 
 <iframe id="yourls-frame" frameborder="0" noresize="noresize" src="$url" name="yourlsFrame"></iframe>
 <script type="text/javascript" src="$pluginurl/js/toolbar.js"></script>
-<script type="text/javascript" src="http://cdn.topsy.com/topsy.js?init=topsyWidgetCreator"></script>
-<script type="text/javascript" src="http://feeds.delicious.com/v2/json/urlinfo/$md5?callback=yourls_get_books"></script>
 </body>
 </html>
 PAGE;
-	
+
 	// Don't forget to die, to interrupt the flow of normal events (ie redirecting to long URL)
 	die();
 }
